% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansPoi.R
\name{kmeansPoi}
\alias{kmeansPoi}
\title{Main Function: k-means clustering for count data}
\usage{
kmeansPoi(data, k, min_mu = 1e-04, verbose = FALSE)
}
\description{
Main Function: k-means clustering for count data
}
